Made get_full_path safer
Running get_full_path for a non-existant file or folder caused it to basically return "$(pwd)/$(basename $1)" which proved confusing to debug. Changed it so it returns a clearly broken path and writes warning to std_err.

Not sure if we should should exit if argument is non-existant or just output the warning. I selected this method to avoid breaking any existing usage.